+++ tests/src/ContainerTest.php
@@ -1122,6 +1122,7 @@ class ContainerTest extends TestCase
         $container->set(HooksCollectionInterface::class, $this->createMock(HooksCollectionInterface::class));
         $container->set('teknoo.east.paas.root_dir', '/foo');
         $container->set('teknoo.east.paas.default_storage_provider', 'foo');
+        $container->set('teknoo.east.paas.default_storage_size', 'foo');
 
         $container->set(
             'teknoo.east.paas.conductor.images_library',